Refactored StandardHeader, moved components, updated imports
