+++ assets/Components.tsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components/native';
 import { SafeAreaView } from 'react-native-safe-area-context';
+import { StyleSheet } from 'react-native';
 import Colors from './Colors';
 
 export const SafeArea = styled(SafeAreaView)`
@@ -29,6 +30,14 @@ export const ButtonMagenta = styled(ButtonBase)`
 export const ButtonWhite = styled(ButtonBase)`
   background: ${Colors.offWhite};
   border-color: ${Colors.magenta};
+`;
+
+export const TitleContainer = styled.View`
+  padding-top: 20px;
+  padding-bottom: 20px;
+  margin-left: 32px;
+  margin-right: 32px;
+  justify-content: center;
 `;
 
 export const Row = styled.View`
@@ -91,10 +100,34 @@ export const CardContainer = styled.View`
   flex: 1;
 `;
 
-export const TitleContainer = styled.View`
-  padding-top: 20px;
-  padding-bottom: 20px;
-  margin-left: 32px;
-  margin-right: 32px;
-  justify-content: center;
-`;
+export const styles = StyleSheet.create({
+  fieldDefault: {
+    borderWidth: 1,
+    height: 35,
+    paddingVertical: 7,
+    paddingHorizontal: 10,
+    marginTop: '3%',
+    marginBottom: '4%',
+    borderRadius: 5,
+    borderColor: Colors.midGray,
+    backgroundColor: Colors.lightGray,
+  },
+  isFocused: {
+    borderColor: Colors.magenta,
+    backgroundColor: Colors.offWhite,
+  },
+  isInvalid: {
+    borderColor: Colors.alertRed,
+    backgroundColor: Colors.alertLightRed,
+  },
+});
+
+export const fieldFocused = StyleSheet.compose(
+  styles.fieldDefault,
+  styles.isFocused,
+);
+
+export const fieldIsInvalid = StyleSheet.compose(
+  styles.fieldDefault,
+  styles.isInvalid,
+);
